CustomizedFoodPage now retrieve the ingredients data from the database, not Edamam API
