+++ .env.sample
@@ -1,4 +1,2 @@
-REACT_APP_API_ID=YOUR_EDAMAM_API_ID
-REACT_APP_SECRET_API_KEY=YOUR_EDAMAM_API_KEY
 SKIP_PREFLIGHT_CHECK=true
 REACT_APP_API_BASE_URL=http://localhost:5000
