+++ src/screens/OrdersRecap.js
@@ -17,7 +17,6 @@ export default function OrdersRecap() {
   useEffect(() => {
     const source = CancelToken.source();
     setLoadingOrders(true);
-    console.log('loadingOrders  ', loadingOrders);
     API.get('/orders', { cancelToken: source.token })
       .then((res) => setOrders(res.data))
       .catch(handleError)
@@ -34,10 +33,6 @@ export default function OrdersRecap() {
       source.cancel('request cancelled');
     };
   }, []);
-
-  console.log(error);
-  console.log(loadingOrders);
-  console.log(orders);
 
   return (
     <div>
